Don't test nonce for truthiness before inclusion
As of version 4.0 of django-csp, the behavior of the
object that wraps the nonce in the request and/or context
has changed.  Testing for truthiness now returns `False`
instead of `True`.

The truthiness check is unnecessary anyway, skipping it
is compatible with all versions of django-csp.  Probably
more performant too, template rendering wise. Not that
it matters.